impl(internal/parser): move code
At the moment package language lives under internal/genclient, which
adds a level of nesting that isn't necessary. Move this up a level.
