+++ generator/internal/parser/openapi/openapi.go
