+++ generator/internal/parser/protobuf/annotations.go
