+++ generator/internal/parser/protobuf/testdata/enum.proto
