+++ generator/internal/parser/protobuf/testdata/imported.proto
