+++ generator/internal/parser/protobuf/testdata/object_fields.proto
