+++ generator/internal/parser/protobuf/testdata/oneofs.proto
