+++ generator/internal/parser/protobuf/testdata/scalar.proto
