+++ generator/internal/parser/protobuf/testdata/scalar_array.proto
