+++ generator/internal/parser/protobuf/testdata/scalar_optional.proto
