+++ generator/internal/parser/protobuf/testdata/with_import.proto
