+++ generator/internal/parser/protobuf/testdata/wkt_fields.proto
